backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
aiosqlite>=0.20.0
httpx>=0.27.0
python-dotenv>=1.0.0

# LLM プロバイダー（使うものだけインストールすれば OK）
anthropic>=0.84.0          # Anthropic (Claude Haiku)
openai>=1.30.0             # OpenAI (GPT-4o-mini) / OpenRouter 共用
google-generativeai>=0.7.0 # Gemini
